add blocks and aspp
