+++ simrec/layers/__init__.py
@@ -1,0 +1,1 @@
+from .conv_layer import vgg_conv, darknet_conv, darknetblock, aspp_decoder
