+++ simrec/layers/blocks.py
@@ -13,13 +13,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import torch
 import torch.nn as nn
 
 
 def vgg_conv(in_ch, out_ch, ksize, stride):
     """
-    Add a vgg conv block.
+    Build a VGG-style convolution block.
+    
     Args:
         in_ch (int): number of input channels of the convolution layer.
         out_ch (int): number of output channels of the convolution layer.
@@ -36,14 +36,18 @@ def vgg_conv(in_ch, out_ch, ksize, stride):
     stage.add_module('relu', nn.ReLU())
     return stage
 
-def darknet_conv(in_ch, out_ch, ksize, stride=1,dilation_rate=1):
+
+def darknet_conv(in_ch, out_ch, ksize, stride=1, dilation_rate=1):
     """
-    Add a conv2d / batchnorm / leaky ReLU block.
+    Add a darknet-style convolution block as Conv-Bn-LeakyReLU.
+    
     Args:
         in_ch (int): number of input channels of the convolution layer.
         out_ch (int): number of output channels of the convolution layer.
         ksize (int): kernel size of the convolution layer.
         stride (int): stride of the convolution layer.
+        dilation_rate (int): spacing between kernel elements.
+    
     Returns:
         stage (Sequential) : Sequential layers composing a convolution block.
     """
@@ -61,6 +65,7 @@ class darknetblock(nn.Module):
     """
     Sequential residual blocks each of which consists of \
     two convolution layers.
+    
     Args:
         ch (int): number of input and output channels.
         nblocks (int): number of residual blocks.
@@ -83,25 +88,4 @@ class darknetblock(nn.Module):
             for res in module:
                 h = res(h)
             x = x + h if self.shortcut else h
-        return x
-
-class aspp_decoder(nn.Module):
-    def __init__(self, planes,hidden_planes,out_planes):
-        super().__init__()
-        self.conv0 = darknet_conv(planes, hidden_planes, ksize=1, stride=1)
-        self.conv1 = darknet_conv(planes, hidden_planes, ksize=3, stride=1,dilation_rate=6)
-        self.conv2 = darknet_conv(planes, hidden_planes, ksize=3, stride=1,dilation_rate=12)
-        self.conv3 = darknet_conv(planes, hidden_planes, ksize=3, stride=1,dilation_rate=18)
-        self.conv4 = darknet_conv(planes, hidden_planes, ksize=1, stride=1)
-        self.pool=nn.AdaptiveAvgPool2d(1)
-        self.out_proj= nn.Conv2d(hidden_planes*5,out_planes,1)
-    def forward(self, x):
-        b,c,h,w=x.size()
-        b0=self.conv0(x)
-        b1=self.conv1(x)
-        b2=self.conv2(x)
-        b3=self.conv3(x)
-        b4=self.conv4(self.pool(x)).repeat(1,1,h,w)
-        x=torch.cat([b0,b1,b2,b3,b4],1)
-        x=self.out_proj(x)
         return x
